minor lint fixes
