+++ shabadkosh/package.json
@@ -72,6 +72,7 @@
     "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-n": "^15.0.0",
     "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.32.2",
     "file-loader": "^6.2.0",
     "prettier": "^2.8.4",
     "source-map-loader": "^4.0.1",
