+++ shabadkosh/src/types/user.ts
@@ -1,4 +1,6 @@
+import { User } from "firebase/auth";
 import { TimestampType } from "./timestamp";
+import { ReactNode } from "react";
 
 export interface NewUserType {
     id?: string,
@@ -10,4 +12,24 @@ export interface NewUserType {
     created_at: TimestampType,
     created_by: string,
     updated_at: TimestampType,
+}
+
+export interface LocalUser {
+    user: User | null,
+    uid?: string,
+    email?: string,
+    displayName?: string,
+    photoURL?: string
+    role?: string,
+    username?: string
+}
+
+export interface ChildrenProps {
+    children?: ReactNode
+}
+
+export type UserState = {
+    auth: {
+      user: LocalUser
+    }
 }
